Text.XML.Expat.SAX

Description

This module provides functions to parse an XML document to a lazy stream of SAX events.

XML primitives

data Encoding

Constructors

ASCII

UTF8

UTF16

ISO88591

data XMLParseError

Parse error, consisting of message text and error location

Constructors

XMLParseError String XMLParseLocation

data XMLParseLocation

Specifies a location of an event within the input text

Constructors

XMLParseLocation

Fields

xmlLineNumber :: Int64

Line number of the event

xmlColumnNumber :: Int64

Column number of the event

xmlByteIndex :: Int64

Byte index of event from start of document

xmlByteCount :: Int64

The number of bytes in the event

SAX-style parse

data ParseOptions tag text

Constructors

ParseOptions

Fields

overrideEncoding :: Maybe Encoding

The encoding parameter, if provided, overrides the document's encoding declaration.

entityDecoder :: Maybe (tag -> Maybe text)

If provided, entity references (i.e. &nbsp; and friends) will be decoded into text using the supplied lookup function

data SAXEvent tag text

Constructors

XMLDeclaration text (Maybe text) (Maybe Bool)

StartElement tag [(tag, text)]

EndElement tag

CharacterData text

StartCData

EndCData

ProcessingInstruction text text

Comment text

FailDocument XMLParseError

parse

Arguments

:: (GenericXMLString tag, GenericXMLString text)

=> ParseOptions tag text
Parse options

-> ByteString
Input text (a lazy ByteString)

-> [SAXEvent tag text]

Lazily parse XML to SAX events. In the event of an error, FailDocument is the last element of the output list.

parseG

Arguments

:: forall tag text l. (GenericXMLString tag, GenericXMLString text, List l)

=> ParseOptions tag text
Parse options

-> l ByteString
Input text (a lazy ByteString)

-> l (SAXEvent tag text)

Parse a generalized list of ByteStrings containing XML to SAX events. In the event of an error, FailDocument is the last element of the output list.

parseLocations

Arguments

:: (GenericXMLString tag, GenericXMLString text)

=> ParseOptions tag text
Parse options

-> ByteString
Input text (a lazy ByteString)

-> [(SAXEvent tag text, XMLParseLocation)]

A variant of parseSAX that gives a document location with each SAX event.

parseLocationsG

Arguments

:: forall tag text l. (GenericXMLString tag, GenericXMLString text, List l)

=> ParseOptions tag text
Parse options

-> l ByteString
Input text (a lazy ByteString)

-> l (SAXEvent tag text, XMLParseLocation)

Parse a generalized list of ByteStrings containing XML to SAX events. In the event of an error, FailDocument is the last element of the output list.

parseLocationsThrowing

Arguments

:: (GenericXMLString tag, GenericXMLString text)

=> ParseOptions tag text
Optional encoding override

-> ByteString
Input text (a lazy ByteString)

-> [(SAXEvent tag text, XMLParseLocation)]

A variant of parseSAX that gives a document location with each SAX event. In the event of an error, throw XMLParseException.

parseLocationsThrowing can throw an exception from pure code, which is generally a bad way to handle errors, because Haskell's lazy evaluation means it's hard to predict where it will be thrown from. However, it may be acceptable in situations where it's not expected during normal operation, depending on the design of your program.

parseThrowing

Arguments

:: (GenericXMLString tag, GenericXMLString text)

=> ParseOptions tag text
Parse options

-> ByteString
input text (a lazy ByteString)

-> [SAXEvent tag text]

Lazily parse XML to SAX events. In the event of an error, throw XMLParseException.

parseThrowing can throw an exception from pure code, which is generally a bad way to handle errors, because Haskell's lazy evaluation means it's hard to predict where it will be thrown from. However, it may be acceptable in situations where it's not expected during normal operation, depending on the design of your program.

defaultParseOptions :: ParseOptions tag text

Variants that throw exceptions

data XMLParseException

An exception indicating an XML parse error, used by the ..Throwing variants.

Constructors

XMLParseException XMLParseError

Abstraction of string types

class (Monoid s, Eq s) => GenericXMLString s where

An abstraction for any string type you want to use as xml text (that is, attribute values or element text content). If you want to use a new string type with hexpat, you must make it an instance of GenericXMLString.

Methods

gxNullString :: s -> Bool

gxToString :: s -> String

gxFromString :: String -> s

gxFromChar :: Char -> s

gxHead :: s -> Char

gxTail :: s -> s

gxBreakOn :: Char -> s -> (s, s)

gxFromByteString :: ByteString -> s

gxToByteString :: s -> ByteString