Text.XML.Expat.Internal.IO

Description

Low-level interface to Expat. Unless speed is paramount, this should normally be avoided in favour of the interfaces provided by SAX and Tree, etc.

Documentation

type HParser = ByteString -> Bool -> IO (ForeignPtr Word8, CInt, Maybe XMLParseError)

hexpatNewParser

Arguments

:: Maybe Encoding

-> Maybe (ByteString -> Maybe ByteString)
Entity decoder

-> Bool
Whether to include input locations

-> IO (HParser, IO XMLParseLocation)

encodingToString :: Encoding -> String

data Encoding

Constructors

ASCII

UTF8

UTF16

ISO88591

data XMLParseError

Parse error, consisting of message text and error location

Constructors

XMLParseError String XMLParseLocation

data XMLParseLocation

Specifies a location of an event within the input text

Constructors

XMLParseLocation

Fields

xmlLineNumber :: Int64

Line number of the event

xmlColumnNumber :: Int64

Column number of the event

xmlByteIndex :: Int64

Byte index of event from start of document

xmlByteCount :: Int64

The number of bytes in the event