Text.XML.Expat.Internal.Namespaced

Documentation

data NName text

A namespace-qualified tag.

NName has two components, a local part and an optional namespace. The local part is the name of the tag. The namespace is the URI identifying collections of declared tags. Tags with the same local part but from different namespaces are distinct. Unqualified tags are those with no namespace. They are in the default namespace, and all uses of an unqualified tag are equivalent.

Constructors

NName

Fields

nnNamespace :: Maybe text

nnLocalPart :: !text

type NAttributes text = Attributes (NName text) text

Type shortcut for attributes with namespaced names

mkNName :: text -> text -> NName text

Make a new NName from a prefix and localPart.

mkAnNName :: text -> NName text

Make a new NName with no prefix.

toNamespaced :: (NodeClass n c, GenericXMLString text, Ord text, Show text) => n c (QName text) text -> n c (NName text) text

fromNamespaced :: (NodeClass n c, GenericXMLString text, Ord text, Functor c) => n c (NName text) text -> n c (QName text) text

xmlnsUri :: GenericXMLString text => text

xmlns :: GenericXMLString text => text